# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RagInt Backend API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "paths": {
...
    "/api/tour/control": {
      "get": {
        "summary": "Get tour control state + commands",
        "parameters": [
          { "name": "since_id", "in": "query", "required": false, "schema": { "type": "integer", "default": 0 } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 50 } },
          { "name": "X-Client-ID", "in": "header", "required": false, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "summary": "Send a tour control command",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["action"],
                "properties": {
                  "action": { "type": "string" },
                  "payload": { "type": "object" },
                  "client_id": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/tour/control/consume": {
      "post": {
        "summary": "Consume a tour control command",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["command_id"],
                "properties": {
                  "command_id": { "type": "integer" },
                  "client_id": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/api/status": {
      "get": {
        "summary": "Request status + timings",
        "parameters": [
          { "name": "request_id", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
    "/api/events": {
      "get": {
        "summary": "Event timeline",
        "parameters": [
          { "name": "request_id", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer" } },
          { "name": "since_ms", "in": "query", "required": false, "schema": { "type": "integer" } },
          { "name": "format", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
...
    "/api/diagnostics": {
      "get": {
        "summary": "Download diagnostics bundle (zip)",
        "parameters": [
          { "name": "key", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "ZIP bundle" },
          "403": { "description": "Forbidden" }
        }
      }
    }
  }
}